moving board stuff into one package with rest of the model
