--- src/main/java/org/sehes/tetris/model/board/IBoardView.java
@@ -1,11 +1,0 @@
-package org.sehes.tetris.model.board;
-
-public interface IBoardView {
-
-    int getWidth();
-
-    int getHeight();
-
-    BlockContent getBlockContent(int row, int column);
-
-}
